Simpler RequestBuilder
