+++ chains/base.json
@@ -2,7 +2,7 @@
     "chains": [
         {
             "name": "juno",            
-            "chain-id": "localjuno-1",
+            "chain_id": "localjuno-1",
             "denom": "ujuno",
             "binary": "junod",
             "bech32-prefix": "juno",
@@ -40,13 +40,13 @@
                 "accounts": [
                     {
                         "name": "acc0",
-                        "address": "juno1efd63aw40lxf3n4mhf7dzhjkr453axurv2zdzk",
+                        "address": "juno1hj5fveer5cjtn4wd6wstzugjfdxzl0xps73ftl", 
                         "amount": "10000000000%DENOM%",
                         "mnemonic": "decorate bright ozone fork gallery riot bus exhaust worth way bone indoor calm squirrel merry zero scheme cotton until shop any excess stage laundry"
                     },
                     {
                         "name": "acc1",
-                        "address": "juno1hj5fveer5cjtn4wd6wstzugjfdxzl0xps73ftl",
+                        "address": "juno1efd63aw40lxf3n4mhf7dzhjkr453axurv2zdzk",
                         "amount": "10000000000%DENOM%",
                         "mnemonic": "wealth flavor believe regret funny network recall kiss grape useless pepper cram hint member few certain unveil rather brick bargain curious require crowd raise"
                     }
