+++ interchain/handlers/actions.go
@@ -25,7 +25,7 @@ func NewActions(ctx context.Context, vals map[string]*cosmos.ChainNode) *actions
 }
 
 type ActionHandler struct {
-	ChainId string `json:"chain-id"`
+	ChainId string `json:"chain_id"`
 	Action  string `json:"action"`
 	Cmd     string `json:"cmd"`
 }
@@ -41,7 +41,7 @@ func (a *actions) PostActions(w http.ResponseWriter, r *http.Request) {
 	chainId := ah.ChainId
 	action := ah.Action
 	if _, ok := a.vals[chainId]; !ok {
-		util.Write(w, []byte(fmt.Sprintf(`{"error":"chain-id %s not found"}`, chainId)))
+		util.Write(w, []byte(fmt.Sprintf(`{"error":"chain_id '%s' not found. Chains %v"}`, chainId, a.vals[chainId])))
 		return
 	}
