+++ scripts/ref_types/__init__.py
@@ -1,0 +1,4 @@
+from .cosmwasm import CosmWasm
+
+# TODO: specific
+from .transactions import *
