+++ scripts/util_base.py
@@ -1,3 +1,4 @@
+import hashlib
 import json
 import os
 
@@ -18,6 +19,59 @@ def default_contracts_json():
 default_contracts_json()
 
 
+# TODO: Do these make sense here?
+def get_chain_start_time_from_logs() -> int:
+    logs_path = os.path.join(parent_dir, "configs", "logs.json")
+
+    with open(logs_path, "r") as f:
+        logs = dict(json.load(f))
+
+    return int(logs.get("start-time", -1))
+
+
+def get_cache_or_default(contracts: dict, ictest_chain_start: int) -> dict:
+    with open(contracts_json_path, "r") as f:
+        cache_time = dict(json.load(f)).get("start_time", 0)
+
+    if cache_time == 0 or cache_time != ictest_chain_start:
+        # reset cache, and set cache time to current ictest time
+        contracts["start_time"] = ictest_chain_start
+        contracts["file_cache"] = {}
+
+        # write to file
+        with open(contracts_json_path, "w") as f:
+            json.dump(contracts, f, indent=4)
+
+    return contracts
+
+
+def update_cache(contracts: dict, code_id: str | int, sha_hash: str) -> int:
+    contracts["file_cache"][sha_hash] = int(code_id)
+    with open(contracts_json_path, "w") as f:
+        json.dump(contracts, f, indent=4)
+    return int(code_id)
+
+
+def get_file_hash(rel_file_path: str, chainId: str) -> str:
+    BUF_SIZE = 65536  # 64k chunks
+    sha1 = hashlib.sha1()
+
+    file_path = os.path.join(contracts_path, rel_file_path)
+
+    if not os.path.exists(file_path):
+        raise FileNotFoundError(f"File not found: {file_path}")
+
+    sha1.update(chainId.replace("-", "").encode("utf-8"))
+    with open(file_path, "rb") as f:
+        while True:
+            data = f.read(BUF_SIZE)
+            if not data:
+                break
+            sha1.update(data)
+
+    return sha1.hexdigest()
+
+
 # create contracts folder if not already
 if not os.path.exists(contracts_path):
     os.mkdir(contracts_path)
@@ -28,5 +82,5 @@ with open(os.path.join(parent_dir, "configs", "server.json")) as f:
 
 PORT = server_config["port"]
 HOST = server_config["host"]
-URL = f"http://{HOST}:{PORT}/"
-URL_UPLOAD = f"http://{HOST}:{PORT}/upload"
+
+API_URL = f"http://{HOST}:{PORT}"
